// src/schema/squirrel.schema.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * JSON schema for configuring squirrel
 */
export interface SquirrelPackagingConfig {
    $schema?: string
    /**
     * Destination for production files
     */
    dest?: string
    /**
     * Entry file name
     */
    entryFileName?: string
    /**
     * Root dir name of the source files
     */
    srcRoot?: string
    /**
     * Relative path to the ts config file
     */
    tsConfig?: string
    /**
     * Files should copied to dest folder
     */
    copyFiles?: string[]
}